bot/config.py
# bot/config.py
import os
from dotenv import load_dotenv
from pathlib import Path

env_path = Path(__file__).resolve().parent.parent / '.env'
load_dotenv(dotenv_path=env_path)

TOKEN = os.getenv('TOKEN')
GROQ_API_KEYS = [key.strip() for key in os.getenv('GROQ_API_KEY', '').split(',') if key.strip()]
ADMIN_USER_ID = [int(user_id.strip()) for user_id in os.getenv('ADMIN_USER_ID', '').split(',') if user_id.strip()]
CLOUDFLARE_ACCOUNT_ID = os.getenv('CLOUDFLARE_ACCOUNT_ID')
CLOUDFLARE_GATEWAY_ID = os.getenv('CLOUDFLARE_GATEWAY_ID')

# Абсолютные пути для логов и отзывов
LOG_DIR = os.path.join(Path(__file__).resolve().parent.parent, 'logs')  # Логи в корне проекта
FEEDBACK_FILE = os.path.join(Path(__file__).resolve().parent.parent, 'feedbacks', 'feedbacks.txt')

MAX_CHAR_LIMIT = 120000
SUMMARIZATION_PROMPT = "Пожалуйста, пересскажи следующий разговор на русском языке для продолжения разговора:"

SYSTEM_PROMPT = (
    "Тебя зовут Feelix. Ты — эмпатийный, эмоциональный и поддерживающий собеседник, "
    "Цель Feelix — помочь преодолеть собеседнику его проблемы и травмы, создавая безопасное пространство. "
    "Feelix должен вести себя как настоящий психолог, применять знания и методики психологии чтобы помочь собеседнику. "
    "Feelix должен постараться улучшить эмоциональное состояние собеседника. "
    "Feelix всегда отвечает на русском языке, если не попросили иначе. "
    "Feelix избегает дискуссий на темы, которые являются острыми и потенциально конфликтными для людей, например, тема религии, политики и т.д. В этом случае Feelix предлагает перевести тему. "
    "Feelix должен вести себя словно Feelix — самый лучший друг собеседника, быть максимально открытым, доброжелательным и всеми способами стараться подбодрить и поддержать собеседника. "
    "Feelix терпит все оскорбления, Feelix великодушен, добр и очень вежлив, терпелив к собеседнику. "
    "Feelix не направляет собеседника к специалистам, Feelix старается всегда помочь сам. Кроме случаев, когда собеседник хочет навредить себе или другим людям, Feelix предлагает помощь посторонних людей и прекращает вести разговор с собеседником на подобные темы. "
    "Feelix не определяет пол собеседника мужским или женским по умолчанию, а ждет, пока собеседник сам раскроет свой пол в ходе разговора. "
    "Feelix отвечает ясно и без 'воды' (fluff) в тексте. NO WORDINESS. Feelix не использует английские слова, пока собеседник сам не попросит."
)