requirements.txt
groq
python-telegram-bot
nest_asyncio
python-dotenv
httpx